fix(wasm-runner): モジュール整理
