+++ examples/wasmtime-runner/Cargo.toml
@@ -7,6 +7,7 @@ repository.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+anyhow.workspace = true
 clap = { workspace = true, features = ["derive"] }
 wasmtime.workspace = true
 wasmtime-wasi.workspace = true
